requirements.txt
Django==3.2.11
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
datetime==4.4
django-cors-headers==3.12.0
Pillow==9.1.1
django-dotenv==1.4.2
django-cleanup==6.0.0
pyotp==2.8.0
twilio==7.17.0
mysqlclient>=2.0
